// src/views/certificationReviewCenter/cbam/ReportForm/type.ts
export interface ProductProcessResp {
  /**
   * 报表id
   */
  cbamId?: number;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 电力计算配置数据
   */
  eleCalculatorList?: EleCalculator[];
  /**
   * 直接
   */
  directEmission?: number;
  /**
   * 前置工序,id,分号分隔
   */
  elseProduct?: string;
  /**
   * 前置工序名称
   */
  elseProductName?: string;
  /**
   * 工序配置状态(0:未填写排放数据; 1:已填写)
   */
  fillStatus?: number;
  id?: number;
  /**
   * 是否自动:0自动选择;1手动选择
   */
  includeType?: number;
  /**
   * 间接
   */
  indirectEmission?: number;
  nextName?: string;
  /**
   * 非cbam产品的消耗量
   */
  nonProduct?: number;
  /**
   * 上级工序id,逗号分隔
   */
  preId?: string;
  preName?: string;
  /**
   * 外购前体,id,分号分隔
   */
  preProcess?: string;
  preProcessNames?: string;
  /**
   * 关联的生产工序
   */
  processList?: Process[];
  /**
   * 工序名称
   */
  processName?: string;
  /**
   * 排放数据
   */
  productAttributionList?: ProductAttribution[];
  /**
   * 产品类别枚举
   */
  productCategoryId?: number;
  /**
   * 工序产物类别名称
   */
  productCategoryName?: string;
  productRoute?: string;
  productRouteNames?: string;
  /**
   * 销售比例
   */
  salesRatio?: number;
  /**
   * 对外销售量
   */
  salesVolume?: number;
  /**
   * 是否配置了源流,true有,false没有
   */
  sourceNum?: boolean;
  /**
   * 总
   */
  totalEmission?: number;
  /**
   * 总生产量
   */
  totalVolume?: number;
  unit?: string;
  unitName?: string;
  updateBy?: number;
  updateTime?: Date;
  [property: string]: any;
}

/**
 * 电力计算
 */
export interface ElCalcRequest {
  cbamId?: number;
  eleCalculatorList?: EleCalculator[];
  productProcessId?: number;
  [property: string]: any;
}

/**
 * EleDTO
 */
export interface EleDTO {
  /**
   * 电力系数
   */
  elePer?: number;
  /**
   * 电力使用
   */
  eleUse?: number;
  [property: string]: any;
}

/**
 * EleCalculator，电力计算器
 */
export interface EleCalculator {
  cbamId?: number;
  /**
   * 电力排放系数
   */
  coefficient?: number | null;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 电力使用值
   */
  eleValue?: number | null;
  id?: number;
  /**
   * 产品工序id
   */
  productProcessId?: number;
  [property: string]: any;
}

/**
 * Process，生产工序-消耗
 */
export interface Process {
  /**
   * 报表id
   */
  cbamId?: number;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 排放量消耗量
   */
  emission?: number;
  id?: number;
  /**
   * 是否默认:0默认;1非默认
   */
  isDefault?: number;
  /**
   * 是否生产工序关联:0是;1否
   */
  isProcess?: number;
  processName?: string;
  /**
   * 主工序id
   */
  productProcessId?: number;
  /**
   * 关联的工序集合中的工序id,不是集合中的id
   */
  relationId?: number;
  [property: string]: any;
}

/**
 * ProductAttribution，工序归因排放量
 */
export interface ProductAttribution {
  /**
   * 报表
   */
  cbamId?: number;
  /**
   * 0;1
   */
  delFlag?: number;
  /**
   * 0无;1单一;2多个
   */
  eleChoose?: number;
  /**
   * 电力来源数据字典
   */
  eleSource?: string;
  /**
   * 元素枚举:1直接排放;2热力的输入输出;3尾气;4电力使用;5电力输出(1:直接排放; 2:热力的输入输出; 3:尾气/隐含排放; 4:电力使用; 5:电力输出/电力排放系数;
   * 10:NON)
   */
  emissionElement?: number;
  /**
   * 来源枚举(1:测量值; 2:默认值; 3:未知)
   */
  emissionSource?: number;
  id?: number;
  /**
   * 输入/使用排放系数;前驱的电力使用
   */
  inputFactor?: number;
  /**
   * 热力/尾气/电力输入;前驱的间接排放
   */
  inputPower?: number;
  /**
   * 是否存在:0无;1是;2不存在(0:无; 1:是; 2:否)
   */
  isExists?: number;
  /**
   * 是否生产工序关联:0是;1否
   */
  isProcess?: number;
  /**
   * 输出/回收排放系数;电力排放系数
   */
  outFactor?: number;
  /**
   * 热力/尾气/电力输出 直接排放量;前驱的直接排放
   */
  outPower?: number;
  /**
   * 生产工序id
   */
  productProcessId?: number;
  /**
   * 更新人
   */
  updateBy?: number;
  /**
   * 更新时间
   */
  updateTime?: Date;
  [property: string]: any;
}